fact: move lua files to namespace
